requirements.txt
Django == 3.2.4
djangorestframework==3.12.4
drf-yasg==1.20.0
django-redis==5.0.0
pymysql==1.0.2
djangorestframework-simplejwt == 4.8.0
qiniu == 7.4.1
django-cors-headers==3.8.0
